+++ server/BUILD.bazel
@@ -4,7 +4,6 @@ go_library(
     name = "server",
     srcs = [
         "buffered_read_conn.go",
-        "column.go",
         "conn.go",
         "conn_stmt.go",
         "driver.go",
@@ -53,6 +52,8 @@ go_library(
         "//privilege/conn",
         "//privilege/privileges",
         "//privilege/privileges/ldap",
+        "//server/internal/column",
+        "//server/internal/dump",
         "//server/internal/util",
         "//server/metrics",
         "//session",
@@ -137,7 +138,6 @@ go_test(
     name = "server_test",
     timeout = "moderate",
     srcs = [
-        "column_test.go",
         "conn_stmt_test.go",
         "conn_test.go",
         "driver_tidb_test.go",
@@ -183,6 +183,7 @@ go_test(
         "//parser/mysql",
         "//parser/terror",
         "//planner/core",
+        "//server/internal/column",
         "//server/internal/util",
         "//session",
         "//sessionctx",
