+++ server/conn_test.go
@@ -22,7 +22,6 @@ import (
 	"crypto/x509"
 	"encoding/binary"
 	"fmt"
-	util2 "github.com/pingcap/tidb/server/internal/util"
 	"io"
 	"path/filepath"
 	"strings"
@@ -37,6 +36,7 @@ import (
 	"github.com/pingcap/tidb/extension"
 	"github.com/pingcap/tidb/parser/auth"
 	"github.com/pingcap/tidb/parser/mysql"
+	util2 "github.com/pingcap/tidb/server/internal/util"
 	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/sessionctx/variable"
 	"github.com/pingcap/tidb/store/mockstore"
