+++ server/driver_tidb_test.go
@@ -21,12 +21,13 @@ import (
 	"github.com/pingcap/tidb/parser/charset"
 	"github.com/pingcap/tidb/parser/model"
 	"github.com/pingcap/tidb/parser/mysql"
+	"github.com/pingcap/tidb/server/internal/column"
 	"github.com/pingcap/tidb/types"
 	"github.com/stretchr/testify/require"
 )
 
-func createColumnByTypeAndLen(tp byte, cl uint32) *ColumnInfo {
-	return &ColumnInfo{
+func createColumnByTypeAndLen(tp byte, cl uint32) *column.ColumnInfo {
+	return &column.ColumnInfo{
 		Schema:       "test",
 		Table:        "dual",
 		OrgTable:     "",
