+++ server/internal/column/BUILD.bazel
@@ -1,0 +1,34 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "column",
+    srcs = [
+        "column.go",
+        "result_encoder.go",
+    ],
+    importpath = "github.com/pingcap/tidb/server/internal/column",
+    visibility = ["//server:__subpackages__"],
+    deps = [
+        "//parser/charset",
+        "//parser/mysql",
+        "//server/internal/dump",
+        "//util/logutil",
+        "@org_uber_go_zap//:zap",
+    ],
+)
+
+go_test(
+    name = "column_test",
+    timeout = "short",
+    srcs = [
+        "column_test.go",
+        "result_encoder_test.go",
+    ],
+    embed = [":column"],
+    flaky = True,
+    shard_count = 4,
+    deps = [
+        "//parser/mysql",
+        "@com_github_stretchr_testify//require",
+    ],
+)
