+++ server/internal/dump/BUILD.bazel
@@ -1,0 +1,26 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "dump",
+    srcs = ["dump.go"],
+    importpath = "github.com/pingcap/tidb/server/internal/dump",
+    visibility = ["//server:__subpackages__"],
+    deps = [
+        "//parser/mysql",
+        "//types",
+    ],
+)
+
+go_test(
+    name = "dump_test",
+    timeout = "short",
+    srcs = ["dump_test.go"],
+    embed = [":dump"],
+    flaky = True,
+    shard_count = 3,
+    deps = [
+        "//sessionctx/stmtctx",
+        "//types",
+        "@com_github_stretchr_testify//require",
+    ],
+)
