+++ server/internal/util/BUILD.bazel
@@ -8,11 +8,6 @@ go_library(
     deps = [
         "//config",
         "//parser/charset",
-        "//parser/mysql",
-        "//server",
-        "//types",
-        "//util/logutil",
-        "@org_uber_go_zap//:zap",
     ],
 )
 
@@ -22,10 +17,6 @@ go_test(
     srcs = ["util_test.go"],
     embed = [":util"],
     flaky = True,
-    shard_count = 8,
-    deps = [
-        "//sessionctx/stmtctx",
-        "//types",
-        "@com_github_stretchr_testify//require",
-    ],
+    shard_count = 4,
+    deps = ["@com_github_stretchr_testify//require"],
 )
