+++ server/plan_replayer_test.go
@@ -18,7 +18,6 @@ import (
 	"archive/zip"
 	"bytes"
 	"database/sql"
-	"github.com/pingcap/tidb/server/internal/util"
 	"io"
 	"os"
 	"path/filepath"
@@ -28,6 +27,7 @@ import (
 	"github.com/go-sql-driver/mysql"
 	"github.com/gorilla/mux"
 	"github.com/pingcap/tidb/domain"
+	"github.com/pingcap/tidb/server/internal/util"
 	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/statistics/handle"
 	"github.com/pingcap/tidb/testkit"
