+++ server/stat_test.go
@@ -16,13 +16,13 @@ package server
 
 import (
 	"context"
-	"github.com/pingcap/tidb/server/internal/util"
 	"testing"
 	"time"
 
 	"github.com/pingcap/failpoint"
 	"github.com/pingcap/tidb/domain/infosync"
 	"github.com/pingcap/tidb/keyspace"
+	"github.com/pingcap/tidb/server/internal/util"
 	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/store/mockstore"
 	"github.com/stretchr/testify/require"
