+++ server/util_test.go
@@ -15,19 +15,21 @@
 package server
 
 import (
+	"testing"
+	"time"
+
 	"github.com/pingcap/tidb/parser/charset"
 	"github.com/pingcap/tidb/parser/mysql"
+	"github.com/pingcap/tidb/server/internal/column"
 	"github.com/pingcap/tidb/server/internal/util"
 	"github.com/pingcap/tidb/types"
 	"github.com/pingcap/tidb/util/chunk"
 	"github.com/pingcap/tidb/util/mock"
 	"github.com/stretchr/testify/require"
-	"testing"
-	"time"
 )
 
 func TestDumpTextValue(t *testing.T) {
-	columns := []*ColumnInfo{{
+	columns := []*column.ColumnInfo{{
 		Type:    mysql.TypeLonglong,
 		Decimal: mysql.NotFixedDec,
 	}}
